requirements.txt
pyinstaller
python-docx
pillow
pywin32
pyPDF2